Update install_goify.sh
After testing the bash script on an actual Linux distro some changes had to be made

+++ install_goify.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+echo "Downloading playwright with depenedencies:"
+go run github.com/playwright-community/playwright-go/cmd/playwright@latest install --with-deps
+
 echo "Compiling Goify..."
 go build -o goify ./app
 
@@ -9,19 +12,28 @@ if [ $? -ne 0 ]; then
     exit 1
 fi
 
-echo "Enter the installation directory (default is /opt/goify):"
+echo "Enter the installation directory (default is ~/Goify):"
 read -p "Install Path: " INSTALL_DIR
-INSTALL_DIR=${INSTALL_DIR:-/opt/goify}
+INSTALL_DIR=${INSTALL_DIR:-~/Goify}
 
 echo "Creating installation folder..."
 mkdir -p "$INSTALL_DIR"
+if [ $? -ne 0 ]; then
+    echo "Failed to create directory '$INSTALL_DIR'. Please check your permissions."
+    read -p "Press any key to exit..." -n1 -s
+    exit 1
+fi
 
 echo "Moving goify to $INSTALL_DIR..."
 mv goify "$INSTALL_DIR"
+if [ $? -ne 0 ]; then
+    echo "Failed to move goify to '$INSTALL_DIR'. Please check your permissions."
+    read -p "Press any key to exit..." -n1 -s
+    exit 1
+fi
 
 echo "Adding $INSTALL_DIR to PATH..."
-echo "export PATH=\$PATH:$INSTALL_DIR" >> ~/.bashrc
-source ~/.bashrc
+echo -e "\nexport PATH=\$PATH:$INSTALL_DIR" >> ~/.bashrc
 
 if [ $? -ne 0 ]; then
     echo "Failed to update PATH. You may need to do it manually."
@@ -30,4 +42,6 @@ if [ $? -ne 0 ]; then
 fi
 
 echo "Installation complete. Goify is ready to use from the terminal!"
+echo "If the goify command isn't yet recognized please run 'source ~/.bashrc'"
 read -p "Press any key to exit..." -n1 -s
+source ~/.bashrc
